+++ Craftsman/Builders/GenericRepositoryBuilder.cs
@@ -48,6 +48,8 @@ public static class GenericRepositoryBuilder
             
             return @$"namespace {classNamespace};
 
+using Ardalis.Specification;
+using Ardalis.Specification.EntityFrameworkCore;
 using {domainClassPath.ClassNamespace};
 using {contextClassPath.ClassNamespace};
 using {exceptionsClassPath.ClassNamespace};
@@ -57,7 +59,7 @@ public interface {interfaceName}<TEntity> : {boundaryServiceName}
     where TEntity : BaseEntity
 {{
     IQueryable<TEntity> Query();
-    Task<TEntity> GetByIdOrDefault(Guid id, bool withTracking = true, CancellationToken cancellationToken = default);
+    Task<TEntity?> GetByIdOrDefault(Guid id, bool withTracking = true, CancellationToken cancellationToken = default);
     Task<TEntity> GetById(Guid id, bool withTracking = true, CancellationToken cancellationToken = default);
     Task<bool> Exists(Guid id, CancellationToken cancellationToken = default);
     Task Add(TEntity entity, CancellationToken cancellationToken = default);    
@@ -65,11 +67,18 @@ public interface {interfaceName}<TEntity> : {boundaryServiceName}
     void Update(TEntity entity);
     void Remove(TEntity entity);
     void RemoveRange(IEnumerable<TEntity> entity);
+    Task<List<TEntity>> ListAsync(ISpecification<TEntity> specification, CancellationToken cancellationToken = default);
+    Task<List<TResult>> ListAsync<TResult>(ISpecification<TEntity, TResult> specification, CancellationToken cancellationToken = default);
+    Task<TEntity?> GetByIdOrDefault(ISpecification<TEntity> specification, CancellationToken cancellationToken = default);
+    Task<TResult?> GetByIdOrDefault<TResult>(ISpecification<TEntity, TResult> specification, CancellationToken cancellationToken = default);
+    Task<TEntity> GetById(ISpecification<TEntity> specification, CancellationToken cancellationToken = default);
+    Task<TResult> GetById<TResult>(ISpecification<TEntity, TResult> specification, CancellationToken cancellationToken = default);
 }}
 
 public abstract class {repoName}<TEntity> : {interfaceName}<TEntity> 
     where TEntity : BaseEntity
 {{
+    private readonly SpecificationEvaluator _specificationEvaluator = SpecificationEvaluator.Default;
     private readonly {dbContextName} _dbContext;
 
     protected {repoName}({dbContextName} dbContext)
@@ -131,6 +140,64 @@ public abstract class {repoName}<TEntity> : {interfaceName}<TEntity>
     public virtual void RemoveRange(IEnumerable<TEntity> entities)
     {{
         _dbContext.Set<TEntity>().RemoveRange(entities);
+    }}
+    
+    public virtual async Task<List<TResult>> ListAsync<TResult>(ISpecification<TEntity, TResult> specification, CancellationToken cancellationToken = default)
+    {{
+        var queryResult = await ApplySpecification(specification).ToListAsync(cancellationToken);
+
+        return specification.PostProcessingAction == null 
+            ? queryResult 
+            : specification.PostProcessingAction(queryResult).ToList();
+    }}
+    
+    public virtual async Task<List<TEntity>> ListAsync(ISpecification<TEntity> specification, CancellationToken cancellationToken = default)
+    {{
+        var queryResult = await ApplySpecification(specification).ToListAsync(cancellationToken);
+        
+        return specification.PostProcessingAction == null 
+            ? queryResult 
+            : specification.PostProcessingAction(queryResult).ToList();
+    }}
+    
+    public virtual async Task<TEntity?> GetByIdOrDefault(ISpecification<TEntity> specification, CancellationToken cancellationToken = default)
+    {{
+        return await ApplySpecification(specification).FirstOrDefaultAsync(cancellationToken);
+    }}
+    
+    public virtual async Task<TResult?> GetByIdOrDefault<TResult>(ISpecification<TEntity, TResult> specification, CancellationToken cancellationToken = default)
+    {{
+        return await ApplySpecification(specification).FirstOrDefaultAsync(cancellationToken);
+    }}
+    
+    public virtual async Task<TEntity> GetById(ISpecification<TEntity> specification, CancellationToken cancellationToken = default)
+    {{
+        var entity = await GetByIdOrDefault(specification, cancellationToken);
+        
+        if(entity == null)
+            throw new NotFoundException($""{{typeof(TEntity).Name}} with a query '{{specification.Query}}' was not found."");
+
+        return entity;
+    }}
+    
+    public virtual async Task<TResult> GetById<TResult>(ISpecification<TEntity, TResult> specification, CancellationToken cancellationToken = default)
+    {{
+        var entity = await GetByIdOrDefault(specification, cancellationToken);
+        
+        if(entity == null)
+            throw new NotFoundException($""{{typeof(TEntity).Name}} with a query '{{specification.Query}}' was not found."");
+
+        return entity;
+    }}
+    
+    private IQueryable<TEntity> ApplySpecification(ISpecification<TEntity> specification, bool evaluateCriteriaOnly = false)
+    {{
+        return _specificationEvaluator.GetQuery(_dbContext.Set<TEntity>().AsQueryable(), specification, evaluateCriteriaOnly);
+    }}
+    
+    private IQueryable<TResult> ApplySpecification<TResult>(ISpecification<TEntity, TResult> specification)
+    {{
+        return _specificationEvaluator.GetQuery(_dbContext.Set<TEntity>().AsQueryable(), specification);
     }}
 }}
 ";